requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
pymongo==4.6.1
mysql-connector-python==8.2.0
Flask-PyMongo==2.3.0